Move content-helpers.ts from lib/ -> client/
