+++ packages/shared/src/logic/index.ts
@@ -10,8 +10,8 @@ export * from '@tloncorp/api/lib/noun';
 export * from './markdown';
 export * from '@tloncorp/api/lib/wayfinding';
 export * from '@tloncorp/api/lib/postContent';
-export * from '@tloncorp/api/lib/content-helpers';
 export * from '@tloncorp/api/lib/pinning';
 export * from '@tloncorp/api/client/activity';
 export * from '@tloncorp/api/client/branch';
+export * from '@tloncorp/api/client/content-helpers';
 export * from '@tloncorp/api/client/deeplinks';
